Support reading the Zip64 EOCD
The code this modifies contains a workaround for par files containing
incorrect offsets (due to data like a shell preamble being at the
beginningg of the file), by assuming that there is nothing between the
end of the central directory, and the start of the EOCD.

For Zip64 files, this contained a Zip64 EOCD plus a Zip64 EOCD Locator
(of 76 bytes) which caused the calculated arc_offset to always be wrong
by 76 bytes.

This code does not currently read the EOCD Locator, because the only
case where this is really useful is in multi-disk archives (which this
code does not support) or if the Zip64 EOCD and its Locator are not
adjacent, in which case arc_offset will be calculated incorrectly
anyway.

There are two ways to make arc_offset more robust, which are worth
trying after some discussion:

1. Also try arc_offset=0 (that is, trust the absolute numbers in the
   file) or simply search between 0 and arc_offset for the expected
   signature.
2. Suggest that people stop writing pars with incorrect offsets to begin
   with.  Info-ZIP already includes an option (`-A`) to fix such
   offsets, although it apparently has a bug fixing Zip64 files (see
   commented test in test_zipimport64.py)