Enable support for lzma, bzip2, zstd
Modularize the code and remove some misleading comment that probably was
brought over from the C implementation.  We do an import every time
which should be relatively cheap, but do not do anything fancy with
caching.

Note that zstd needs to be told the uncompressed size, so the functions
themselves are not idempotent.